Initial pass at working XML to VCF converter
This is a simple working convert to move from Microsoft Office Contacts XML
to VCF cards.

We attempt to use the full Vcard RFC as much as possible.

Because the initial source data for conversion used a hacky method for handling
the Christmas Card labels via the "Title" field, we use an X-RELATEDNAME /
X-LABEL field to handle a custom "Christmas Card Name" field.