Apply enhancements written by Codeium
On Wednesday, December 25th, we employed Codeium, leveraging "Cloude 3.5
Sonnet", to suggest enhancements to both `ppxml` and `dumpxmltags`.  We
 reviewed by both sets of code changes and accepted all changes as is.

That, perhaps, is not a good idea, but it was hard to find anything
inherently _wrong_ with the changes.

Further, we instructed Codeium to add usage and documenation information
to the `README.md` for both programs.