Rewrite `dumpxmltags` into `contacttagfreq`
What we really used to work through how to handle each XML tag in a contact was
the frequency table of those tags.  We first reviewed the most frequent tags,
and the moved steadily down through the list in order.

This table enables us to review a contact XML file a bit easier.