Failed attempt to install `poetry` with Codeium
Using the prompt, "Convert this repository to use poetry.", Codeium
created the `contacts` subdirectory, with the `__init__.py` file, moved
the appropriate files there, and then created an initial
`pyproject.toml`.  We filled in the actual user name and email.

However, problems installing `lxml` with `poetry install` caused this
project to be shelved on a branch for now.