Remove indentation option from `ppxml`
Not sure what happened here, but the `etree.toString()` method does
NOT take an option to specify how many spaces to indent lines.  It is
not clear how this slipped in because `lxml` documentation clearly
does not have this documented.

First example of where Codeium got it wrong, and human error (or wacky
environment) without unit tests let it be committed.