Improvements based on import experimentation
We started importing generated `.vcf` output to see how well it works.
The following changes were made based on those experiments.

Report any tags that we didn't find in the source XML.  We want to make
sure that we have processed all the tags, and that if the source input
doesn't cover them somewhere, we have made a mistake in building up the
conversion logic.

When we process *Country tags we don't want to consider any data for the
US, it is assumed.

Turn the big `if`/`continue` construct into `if`/`elif` and drop
`continue`.

The 'company', 'office', 'department' fields does not import well into
Apple's Contacts.  Contacts really only wants "Company" and "Department"
in the `ORG` field.  We combine 'office' and 'department' as
"Department, Office".

Use just `maiden` for maiden names.