Final changes made before import finished
This commit represents the final bit of work needed to finalize the
transformation before actually importaing into Apple's Contacts program.

We added a "Card ID" in the `NOTE` field to debug where things were
failing on partial imports.

This allowed us to find that all the `ADR` fields need to be "backslash-
escaped".

We also added a `PRODID` field.